+++ cli/config.py
@@ -1,5 +1,6 @@
 from pathlib import Path
 
+from mako.lookup import TemplateLookup
 from xdgconfig import JsonConfig
 
 
@@ -12,6 +13,7 @@ class DefaultConfig:
         'app.editor.flags': [],
         'app.pager.name': 'less',
         'app.pager.flags': [],
+        'app.default_verbosity': 0,
     }
 
 
@@ -28,11 +30,15 @@ plugins_path: Path = (config.app_path / 'plugins')
 plugins_path.mkdir(parents=True, exist_ok=True)
 plugins_registry = JsonConfig('cli', 'registry.json')
 state = {
-    'json': False,
-    'csv': False,
-    'toml': False,
-    'yaml': False,
+    'format': 'default',
+    'verbosity': config['app.default_verbosity'],
 }
 RESERVED_COMMANDS = ['config', 'env', 'plugins']
 logs_path: Path = (config.app_path / 'logs')
 logs_path.mkdir(parents=True, exist_ok=True)
+tmp_dir: Path = (config.app_path / 'tmp')
+
+templates = TemplateLookup(
+    directories=[str(templates_path.resolve())],
+    module_directory=str((tmp_dir / 'mako_modules').resolve()),
+)